Add more unit tests for SNI port filtering

+++ iocore/net/YamlSNIConfig.h
@@ -27,6 +27,7 @@
 #include <set>
 #include <optional>
 #include <memory>
+#include <cstdint>
 
 #include "SSLTypes.h"
 
@@ -74,7 +75,7 @@ struct YamlSNIConfig {
 
   struct Item {
     std::string fqdn;
-    std::vector<std::pair<long, long>> port_ranges;
+    std::vector<std::pair<uint16_t, uint16_t>> port_ranges;
     std::optional<bool> offer_h2;   // Has no value by default, so do not initialize!
     std::optional<bool> offer_quic; // Has no value by default, so do not initialize!
     uint8_t verify_client_level = 255;
